refactored kafka tests in order to decrease number of contexts used in tests
