+++ embedded-kafka/src/test/java/com/playtika/test/kafka/EmbeddedKafkaWithoutBindingTest.java
@@ -6,27 +6,23 @@ import org.junit.jupiter.api.AfterAll;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.TestInstance;
 import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.boot.test.context.SpringBootTest;
+import org.springframework.test.context.TestPropertySource;
 
 import java.nio.file.Path;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
-@SpringBootTest(
-        properties = {
-                "embedded.zookeeper.fileSystemBind.enabled=false",
-                "embedded.zookeeper.fileSystemBind.dataFolder=target/embedded-zk-data-without-binding",
-                "embedded.zookeeper.fileSystemBind.txnLogsFolder=target/embedded-zk-txn-logs-without-binding",
+@TestPropertySource(properties = {
+        "embedded.zookeeper.fileSystemBind.enabled=false",
+        "embedded.zookeeper.fileSystemBind.dataFolder=target/embedded-zk-data-without-binding",
+        "embedded.zookeeper.fileSystemBind.txnLogsFolder=target/embedded-zk-txn-logs-without-binding",
 
-                "embedded.kafka.topicsToCreate=autoCreatedTopic",
-                "embedded.kafka.fileSystemBind.enabled=false",
-                "embedded.kafka.fileSystemBind.dataFolder=target/embedded-kafka-data-without-binding"
-        },
-        classes = BaseEmbeddedKafkaTest.TestConfiguration.class
-)
+        "embedded.kafka.fileSystemBind.enabled=false",
+        "embedded.kafka.fileSystemBind.dataFolder=target/embedded-kafka-data-without-binding"
+})
 @TestInstance(TestInstance.Lifecycle.PER_CLASS)
 @DisplayName("Test that embedded-kafka without filesystem binding")
-public class EmbeddedKafkaWithoutBindingTest extends BaseEmbeddedKafkaTest {
+public class EmbeddedKafkaWithoutBindingTest extends EmbeddedKafkaTest {
     @Autowired
     private ZookeeperConfigurationProperties zookeeperProperties;
 
@@ -34,7 +30,7 @@ public class EmbeddedKafkaWithoutBindingTest extends BaseEmbeddedKafkaTest {
     private KafkaConfigurationProperties kafkaProperties;
 
     @AfterAll
-    public void shouldBindToFileSystem() {
+    public void shouldBindToFileSystem() throws Exception {
         Path projectDir = projectDir();
         Path zookeeperDataFolder = projectDir.resolve(zookeeperProperties.getFileSystemBind().getDataFolder());
         Path zookeeperTxnLogsFolder = projectDir.resolve(zookeeperProperties.getFileSystemBind().getTxnLogsFolder());
